+++ linux/flutter/generated_plugin_registrant.cc
@@ -9,7 +9,6 @@
 #include <file_selector_linux/file_selector_plugin.h>
 #include <flutter_secure_storage_linux/flutter_secure_storage_linux_plugin.h>
 #include <smart_auth/smart_auth_plugin.h>
-#include <url_launcher_linux/url_launcher_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) file_selector_linux_registrar =
@@ -21,7 +20,4 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) smart_auth_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "SmartAuthPlugin");
   smart_auth_plugin_register_with_registrar(smart_auth_registrar);
-  g_autoptr(FlPluginRegistrar) url_launcher_linux_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "UrlLauncherPlugin");
-  url_launcher_plugin_register_with_registrar(url_launcher_linux_registrar);
 }
