+++ pubspec.yaml
@@ -19,8 +19,8 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.1.5 <4.0.0"
-  flutter: ">=3.13.9"
+  sdk: ">=3.3.3 <4.0.0"
+  flutter: ">=3.19.5"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,7 +39,7 @@ dependencies:
   image_picker: ^1.0.7
   file_selector: ^1.0.3
 
-  go_router: ^13.2.2
+  go_router: ^13.2.3
 
   bloc: ^8.1.4  
   bloc_concurrency: ^0.2.5
@@ -47,7 +47,7 @@ dependencies:
 
   dio: ^5.4.2+1
   dio_http2_adapter: ^2.5.1
-  alice: ^0.4.1
+  flutter_alice: ^1.1.1
 
   equatable: ^2.0.5
   freezed: ^2.5.1
@@ -68,7 +68,7 @@ dependencies:
   permission_handler: ^11.3.1
   flutter_secure_storage: ^9.0.0
   logger: ^2.2.0
-  connectivity_plus: ^6.0.1
+  connectivity_plus: ^6.0.2
   intl: ^0.19.0
   stream_transform: ^2.1.0
   flavor: ^2.0.0
@@ -101,9 +101,7 @@ dev_dependencies:
 dependency_overrides:
   intl: ^0.19.0
   reactive_forms: ^17.0.0
-#   analyzer: ^5.11.1
-#   path: ^1.8.3
-#   intl: ^0.18.1
+  http: ^1.2.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
