+++ windows/flutter/generated_plugins.cmake
@@ -8,9 +8,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage_windows
   geolocator_windows
   permission_handler_windows
-  share_plus
   smart_auth
-  url_launcher_windows
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
